refactor: scss for header border bottom
