+++ components/Header/Header.module.scss
@@ -1,0 +1,168 @@
+.header {
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+  background-color: #0c0c0c;
+  border-color: #333333;
+  border-top-style: solid;
+  border-top-width: 1px;
+  height: 60px;
+  position: relative;
+  width: 100%;
+  border-bottom: 1px solid #333;
+  min-width: 800px;
+
+  .inctagram {
+    color: #ffffff;
+    font-size: 26px;
+    font-weight: 600;
+    left: 59px;
+    letter-spacing: 0;
+    line-height: 36px;
+    position: absolute;
+    top: 10px;
+    white-space: nowrap;
+  }
+}
+
+.outline_bell {
+  height: 24px;
+  width: 24px;
+  margin-right: 48px;
+  margin-left: 1053px;
+
+  .overlap_group {
+    height: 25px;
+    position: relative;
+    top: -1px;
+
+    .mask {
+      height: 20px;
+      left: 0;
+      position: absolute;
+      top: 5px;
+      width: 18px;
+    }
+
+    .ellipse {
+      background-color: #cc1439;
+      border-radius: 6.5px;
+      height: 13px;
+      left: 11px;
+      position: absolute;
+      top: -4px;
+      width: 13px;
+    }
+
+    .element {
+      color: #ffffff;
+      font-size: 10px;
+      font-weight: 500;
+      height: 13px;
+      left: 11px;
+      letter-spacing: 0;
+      line-height: 5px;
+      position: absolute;
+      text-align: center;
+      top: 0;
+      width: 13px;
+    }
+  }
+}
+
+.label {
+  background-color: transparent;
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  width: 100%;
+
+  .div_wrapper {
+    height: 13px;
+    width: 13px;
+  }
+
+  .text_wrapper {
+    color: #ffffff;
+    font-size: 10px;
+    font-weight: 500;
+    height: 13px;
+    left: 0;
+    letter-spacing: 0;
+    line-height: 8px;
+    position: fixed;
+    text-align: center;
+    top: 0;
+    width: 13px;
+  }
+}
+
+.box {
+  background-color: transparent;
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  width: 100%;
+
+  .ellipse_wrapper {
+    height: 13px;
+    width: 13px;
+
+    .ellipse {
+      background-color: #cc1439;
+      border-radius: 6.5px;
+      height: 13px;
+      left: 0;
+      position: fixed;
+      top: 0;
+      width: 13px;
+    }
+  }
+
+  .mask_wrapper {
+    height: 20px;
+    width: 18px;
+
+    .mask {
+      height: 20px;
+      left: 0;
+      position: fixed;
+      top: 0;
+      width: 18px;
+    }
+  }
+}
+
+.wraper {
+  position: relative;
+
+  .select_container {
+    position: absolute;
+    top: 100%;
+    left: 0;
+    z-index: 1;
+  }
+
+  .language {
+    border: 1px solid grey;
+    background-color: black;
+    background-size: 100% 100%;
+    height: 36px;
+    margin-right: 64px;
+    display: flex;
+    width: 163px;
+    flex-direction: row;
+    justify-content: space-evenly;
+    align-items: center;
+    cursor: pointer;
+
+    .text {
+      color: #ffffff;
+    }
+
+    .text_select {
+      color: #ffffff;
+      padding-right: 40px;
+    }
+  }
+}
